requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-nvidia-ai-endpoints>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
supabase>=2.3.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
sympy>=1.12
matplotlib>=3.8.0
PyMuPDF>=1.24.0
plotly>=5.20.0
pandas>=2.0.0
datasets>=2.14.0